+++ reductstore/src/replication/transaction_log.rs
@@ -1,9 +1,9 @@
-// Copyright 2023-2024 ReductSoftware UG
+// Copyright 2023-2025 ReductSoftware UG
 // Licensed under the Business Source License 1.1
 
 use crate::core::file_cache::FILE_CACHE;
 use crate::replication::Transaction;
-use log::{debug, error, warn};
+use log::{debug, warn};
 use reduct_base::error::ReductError;
 use reduct_base::internal_server_error;
 use std::io::{Read, Seek, SeekFrom, Write};
@@ -28,28 +28,6 @@ pub(super) struct TransactionLog {
 const HEADER_SIZE: usize = 16;
 const ENTRY_SIZE: usize = 9;
 
-impl Drop for TransactionLog {
-    fn drop(&mut self) {
-        // Flush and sync the file in a separate thread.
-        let file = FILE_CACHE.write_or_create(&self.file_path, SeekFrom::Current(0));
-        if file.is_err() {
-            error!("Failed to open transaction log: {}", file.err().unwrap());
-            return;
-        }
-
-        let file = file.unwrap().upgrade().unwrap();
-        let sync = || {
-            let mut file = file.write()?;
-            file.flush()?;
-            file.sync_all()?;
-
-            Ok::<(), ReductError>(())
-        };
-
-        sync().map_or_else(|e| error!("Failed to sync transaction log: {}", e), |r| r);
-    }
-}
-
 impl TransactionLog {
     /// Create a new transaction log or load an existing one.
     ///
@@ -61,7 +39,7 @@ impl TransactionLog {
     /// # Returns
     ///
     /// A new transaction log instance or an error.
-    pub fn try_load_or_create(path: PathBuf, capacity: usize) -> Result<Self, ReductError> {
+    pub fn try_load_or_create(path: &PathBuf, capacity: usize) -> Result<Self, ReductError> {
         let init_capacity_in_bytes = capacity * ENTRY_SIZE + HEADER_SIZE;
 
         let instance = if !path.try_exists()? {
@@ -77,7 +55,7 @@ impl TransactionLog {
             file.sync_all()?;
 
             Self {
-                file_path: path,
+                file_path: path.clone(),
                 capacity_in_bytes: init_capacity_in_bytes,
                 write_pos: HEADER_SIZE,
                 read_pos: HEADER_SIZE,
@@ -94,6 +72,15 @@ impl TransactionLog {
             };
             let write_pos = u64::from_be_bytes(buf[0..8].try_into().unwrap()) as usize;
             let read_pos = u64::from_be_bytes(buf[8..16].try_into().unwrap()) as usize;
+
+            Self::integrity_check(
+                path,
+                init_capacity_in_bytes,
+                capacity_in_bytes,
+                write_pos,
+                read_pos,
+            )?;
+
             let capacity_in_bytes = if init_capacity_in_bytes != capacity_in_bytes {
                 // If the capacity is changed, we need to check if the log is empty
                 // then we can change the capacity.
@@ -118,7 +105,7 @@ impl TransactionLog {
             };
 
             Self {
-                file_path: path,
+                file_path: path.to_path_buf(),
                 capacity_in_bytes,
                 write_pos,
                 read_pos,
@@ -126,6 +113,44 @@ impl TransactionLog {
         };
 
         Ok(instance)
+    }
+
+    fn integrity_check(
+        path: &PathBuf,
+        init_capacity_in_bytes: usize,
+        capacity_in_bytes: usize,
+        write_pos: usize,
+        read_pos: usize,
+    ) -> Result<(), ReductError> {
+        if init_capacity_in_bytes < HEADER_SIZE
+            || (capacity_in_bytes - HEADER_SIZE) % ENTRY_SIZE != 0
+        {
+            return Err(internal_server_error!(
+                "Invalid size {} of transaction log {}",
+                capacity_in_bytes,
+                path.to_str().unwrap_or("unknown path")
+            ));
+        }
+
+        let check_pos = |pos: usize, name: &str| {
+            if pos < HEADER_SIZE
+                || pos >= capacity_in_bytes
+                || (pos - HEADER_SIZE) % ENTRY_SIZE != 0
+            {
+                return Err(internal_server_error!(
+                    "Invalid {} position {} in transaction log {}",
+                    name,
+                    pos,
+                    path.to_str().unwrap_or("unknown path")
+                ));
+            }
+            Ok(())
+        };
+
+        check_pos(write_pos, "write")?;
+        check_pos(read_pos, "read")?;
+
+        Ok(())
     }
 
     /// Push a new transaction to the log.
@@ -271,13 +296,13 @@ mod tests {
 
     #[rstest]
     fn test_new_transaction_log(path: PathBuf) {
-        let transaction_log = TransactionLog::try_load_or_create(path, 100).unwrap();
+        let transaction_log = TransactionLog::try_load_or_create(&path, 100).unwrap();
         assert_eq!(transaction_log.is_empty(), true);
     }
 
     #[rstest]
     fn test_write_read_transaction_log(path: PathBuf) {
-        let mut transaction_log = TransactionLog::try_load_or_create(path, 100).unwrap();
+        let mut transaction_log = TransactionLog::try_load_or_create(&path, 100).unwrap();
         assert_eq!(
             transaction_log
                 .push_back(Transaction::WriteRecord(1))
@@ -308,7 +333,7 @@ mod tests {
 
     #[rstest]
     fn test_write_broken_type(path: PathBuf) {
-        let mut transaction_log = TransactionLog::try_load_or_create(path.clone(), 100).unwrap();
+        let mut transaction_log = TransactionLog::try_load_or_create(&path, 100).unwrap();
         assert_eq!(
             transaction_log
                 .push_back(Transaction::WriteRecord(1))
@@ -337,7 +362,7 @@ mod tests {
 
     #[rstest]
     fn test_out_of_range(path: PathBuf) {
-        let mut transaction_log = TransactionLog::try_load_or_create(path, 100).unwrap();
+        let mut transaction_log = TransactionLog::try_load_or_create(&path, 100).unwrap();
         assert_eq!(
             transaction_log
                 .push_back(Transaction::WriteRecord(1))
@@ -369,7 +394,7 @@ mod tests {
 
     #[rstest]
     fn test_overflow(path: PathBuf) {
-        let mut transaction_log = TransactionLog::try_load_or_create(path, 3).unwrap();
+        let mut transaction_log = TransactionLog::try_load_or_create(&path, 3).unwrap();
         for i in 1..5 {
             transaction_log
                 .push_back(Transaction::WriteRecord(i))
@@ -385,14 +410,14 @@ mod tests {
 
     #[rstest]
     fn test_recovery(path: PathBuf) {
-        let mut transaction_log = TransactionLog::try_load_or_create(path.clone(), 3).unwrap();
+        let mut transaction_log = TransactionLog::try_load_or_create(&path, 3).unwrap();
         for i in 1..5 {
             transaction_log
                 .push_back(Transaction::WriteRecord(i))
                 .unwrap();
         }
 
-        let mut transaction_log = TransactionLog::try_load_or_create(path, 3).unwrap();
+        let mut transaction_log = TransactionLog::try_load_or_create(&path, 3).unwrap();
         assert_eq!(transaction_log.len(), 2);
         assert_eq!(
             transaction_log.front(2).unwrap(),
@@ -405,27 +430,27 @@ mod tests {
 
     #[rstest]
     fn test_recovery_init(path: PathBuf) {
-        let mut transaction_log = TransactionLog::try_load_or_create(path.clone(), 3).unwrap();
+        let mut transaction_log = TransactionLog::try_load_or_create(&path, 3).unwrap();
         transaction_log
             .push_back(Transaction::WriteRecord(1))
             .unwrap();
         drop(transaction_log);
 
-        let transaction_log = TransactionLog::try_load_or_create(path, 3).unwrap();
+        let transaction_log = TransactionLog::try_load_or_create(&path, 3).unwrap();
         assert_eq!(transaction_log.write_pos, HEADER_SIZE + ENTRY_SIZE);
         assert_eq!(transaction_log.read_pos, HEADER_SIZE);
     }
 
     #[rstest]
     fn test_recovery_empty_cache(path: PathBuf) {
-        let mut transaction_log = TransactionLog::try_load_or_create(path.clone(), 3).unwrap();
+        let mut transaction_log = TransactionLog::try_load_or_create(&path, 3).unwrap();
         transaction_log
             .push_back(Transaction::WriteRecord(1))
             .unwrap();
 
         FILE_CACHE.discard_recursive(&path).unwrap(); // discard the cache to simulate restart
 
-        let mut transaction_log = TransactionLog::try_load_or_create(path, 3).unwrap();
+        let mut transaction_log = TransactionLog::try_load_or_create(&path, 3).unwrap();
 
         // check if the transaction log is still working after cache discard
         assert_eq!(
@@ -438,13 +463,13 @@ mod tests {
 
     #[rstest]
     fn test_resize_empty_log(path: PathBuf) {
-        TransactionLog::try_load_or_create(path.clone(), 3).unwrap();
+        TransactionLog::try_load_or_create(&path, 3).unwrap();
         assert_eq!(
             fs::metadata(&path).unwrap().len() as usize,
             ENTRY_SIZE * 3 + HEADER_SIZE
         );
 
-        TransactionLog::try_load_or_create(path.clone(), 5).unwrap();
+        TransactionLog::try_load_or_create(&path, 5).unwrap();
         assert_eq!(
             fs::metadata(&path).unwrap().len() as usize,
             ENTRY_SIZE * 5 + HEADER_SIZE
@@ -453,7 +478,7 @@ mod tests {
 
     #[rstest]
     fn test_resize_non_empty_log(path: PathBuf) {
-        let mut transaction_log = TransactionLog::try_load_or_create(path.clone(), 3).unwrap();
+        let mut transaction_log = TransactionLog::try_load_or_create(&path, 3).unwrap();
         transaction_log
             .push_back(Transaction::WriteRecord(1))
             .unwrap();
@@ -462,12 +487,75 @@ mod tests {
             ENTRY_SIZE * 3 + HEADER_SIZE
         );
 
-        TransactionLog::try_load_or_create(path.clone(), 5).unwrap();
+        TransactionLog::try_load_or_create(&path, 5).unwrap();
         assert_eq!(
             fs::metadata(&path).unwrap().len() as usize,
             ENTRY_SIZE * 3 + HEADER_SIZE,
             "The log is not empty, so the capacity should not be changed."
         );
+    }
+
+    mod integrity_tests {
+        use super::*;
+
+        #[rstest]
+        #[case([0, 0, 0, 0, 0, 0, 0, 17, 0, 0, 0, 0, 0, 0, 0, 16], "Invalid write position 17 in transaction log"
+        )]
+        #[case([0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 16], "Invalid write position 1 in transaction log"
+        )]
+        #[case([0, 0, 0, 0, 0, 0, 0, 34, 0, 0, 0, 0, 0, 0, 0, 16], "Invalid write position 34 in transaction log"
+        )]
+        #[case([0, 0, 0, 0, 0, 0, 0, 16, 0, 0, 0, 0, 0, 0, 0, 17], "Invalid read position 17 in transaction log"
+        )]
+        #[case([0, 0, 0, 0, 0, 0, 0, 16, 0, 0, 0, 0, 0, 0, 0, 1], "Invalid read position 1 in transaction log"
+        )]
+        #[case([0, 0, 0, 0, 0, 0, 0, 16, 0, 0, 0, 0, 0, 0, 0, 34], "Invalid read position 34 in transaction log"
+        )]
+        fn test_invalid_position(
+            #[case] buf: [u8; 16],
+            #[case] expected_error: &str,
+            path: PathBuf,
+        ) {
+            {
+                let file = FILE_CACHE
+                    .write_or_create(&path, SeekFrom::Start(0))
+                    .unwrap()
+                    .upgrade()
+                    .unwrap();
+                let mut file = file.write().unwrap();
+                file.write_all(&buf).unwrap();
+                file.set_len((HEADER_SIZE + ENTRY_SIZE * 2) as u64).unwrap();
+            }
+
+            let result = TransactionLog::try_load_or_create(&path, 3);
+            assert_eq!(
+                result.err().unwrap(),
+                internal_server_error!("{} {}", expected_error, path.to_str().unwrap())
+            );
+        }
+
+        #[rstest]
+        fn test_invalid_size(path: PathBuf) {
+            {
+                let file = FILE_CACHE
+                    .write_or_create(&path, SeekFrom::Start(0))
+                    .unwrap()
+                    .upgrade()
+                    .unwrap();
+                let mut file = file.write().unwrap();
+                file.write_all(&[0u8; 16]).unwrap();
+                file.set_len((HEADER_SIZE + 1) as u64).unwrap();
+            }
+
+            let result = TransactionLog::try_load_or_create(&path, 3);
+            assert_eq!(
+                result.err().unwrap(),
+                internal_server_error!(
+                    "Invalid size 17 of transaction log {}",
+                    path.to_str().unwrap()
+                )
+            );
+        }
     }
 
     #[fixture]